requirements.txt
apispec==6.4.0
apispec-webframeworks==1.0.0
blinker==1.7.0
click==8.1.7
Flask==3.0.0
importlib-metadata==7.0.1
ipdb==0.13.13
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.3
psycopg2==2.9.9
pylint==3.0.3
pytest==7.4.4
python-dotenv==1.0.0
PyYAML==6.0.1
Werkzeug==3.0.1
zipp==3.17.0
